Part 1: Preliminary feature implementation
